refactor: simplify publish token expiry calls
- Pass optional expiry directly to GitHub helpers
- Split oversized publish review tests by responsibility
